requirements.txt
flask
flask-cors
pymongo
python-dotenv
tensorflow
joblib
scikit-learn
nltk
numpy
gunicorn  # For deployment